+++ gamestudio_backend/src/main/java/sk/tuke/gamestudio_backend/service/interfaces/RatingService.java
@@ -5,7 +5,7 @@ import sk.tuke.gamestudio_backend.entity.Rating;
 
 public interface RatingService {
     void setRating(Rating rating) throws RatingException;
-    int getAverageRating(String game) throws RatingException;
+    double getAverageRating(String game) throws RatingException;
     int getRating(String game, String player) throws RatingException;
     void reset() throws RatingException;
 }
