+++ gamestudio_backend/src/main/java/sk/tuke/gamestudio_backend/service/other/RatingServiceJDBC.java
@@ -78,7 +78,7 @@ public class RatingServiceJDBC implements RatingService {
     }
 
     @Override
-    public int getAverageRating(String game) throws RatingException {
+    public double getAverageRating(String game) throws RatingException {
         try(Connection connection = DBConnection.getConnection();
             PreparedStatement preparedStatement = connection.prepareStatement(SELECTAVG)) {
